add render to utils.js file and finish calendar module
