+++ public/utils.js
@@ -7,3 +7,10 @@ export const cleanHtml = htmlString =>htmlString
   .trim()
 
   export const tokenAdmin = () => localStorage.getItem('token')
+
+const mainDiv = document.getElementById('main')
+
+
+export const render = html => {
+    mainDiv.innerHTML = html
+  }
